requirements.txt
tqdm==4.55.1
scipy==1.2.1
wget==3.2
torchvision==0.8.2
# nltk==3.5
matplotlib==3.3.2
h5py==2.8.0
scikit_image==0.17.2
streamlit==0.74.1
pandas
requests==2.25.1
torch==1.7.1
numpy==1.19.2
Pillow==8.1.0
scikit_learn==0.24.1
